Fixed Competitions ordering
Realized the abbreviations of the competitions were causing the ordering issue. LCC's 1-9 were simply labeled "LCC1, LCC2," etc, which conflicted once we got to LCC10 and beyond. So adding leading 0's to those sorted it out. We lucked out thate the later LCC's were years, and the SaberComp years "SC" happen to fall later in the alphabet.